deferred actions: add common functionality for evaluating unexpanded and expanded modules
