+++ internal/terraform/eval_context_builtin.go
@@ -440,7 +440,10 @@ func (ctx *BuiltinEvalContext) EvaluationScope(self addrs.Referenceable, source 
 	switch scope := ctx.scope.(type) {
 	case evalContextModuleInstance:
 		data := &evaluationStateData{
-			Evaluator:       ctx.Evaluator,
+			evaluationData: &evaluationData{
+				Evaluator: ctx.Evaluator,
+				Module:    scope.Addr.Module(),
+			},
 			ModulePath:      scope.Addr,
 			InstanceKeyData: keyData,
 			Operation:       ctx.Evaluator.Operation,
@@ -460,7 +463,10 @@ func (ctx *BuiltinEvalContext) EvaluationScope(self addrs.Referenceable, source 
 		return evalScope
 	case evalContextPartialExpandedModule:
 		data := &evaluationPlaceholderData{
-			Evaluator:      ctx.Evaluator,
+			evaluationData: &evaluationData{
+				Evaluator: ctx.Evaluator,
+				Module:    scope.Addr.Module(),
+			},
 			ModulePath:     scope.Addr,
 			CountAvailable: keyData.CountIndex != cty.NilVal,
 			EachAvailable:  keyData.EachKey != cty.NilVal,
