+++ internal/terraform/evaluate_test.go
@@ -30,7 +30,9 @@ func TestEvaluatorGetTerraformAttr(t *testing.T) {
 		NamedValues: namedvals.NewState(),
 	}
 	data := &evaluationStateData{
-		Evaluator: evaluator,
+		evaluationData: &evaluationData{
+			Evaluator: evaluator,
+		},
 	}
 	scope := evaluator.Scope(data, nil, nil, lang.ExternalFuncs{})
 
@@ -61,7 +63,9 @@ func TestEvaluatorGetPathAttr(t *testing.T) {
 		NamedValues: namedvals.NewState(),
 	}
 	data := &evaluationStateData{
-		Evaluator: evaluator,
+		evaluationData: &evaluationData{
+			Evaluator: evaluator,
+		},
 	}
 	scope := evaluator.Scope(data, nil, nil, lang.ExternalFuncs{})
 
@@ -127,7 +131,9 @@ func TestEvaluatorGetOutputValue(t *testing.T) {
 	}
 
 	data := &evaluationStateData{
-		Evaluator: evaluator,
+		evaluationData: &evaluationData{
+			Evaluator: evaluator,
+		},
 	}
 	scope := evaluator.Scope(data, nil, nil, lang.ExternalFuncs{})
 
@@ -196,7 +202,9 @@ func TestEvaluatorGetInputVariable(t *testing.T) {
 	}
 
 	data := &evaluationStateData{
-		Evaluator: evaluator,
+		evaluationData: &evaluationData{
+			Evaluator: evaluator,
+		},
 	}
 	scope := evaluator.Scope(data, nil, nil, lang.ExternalFuncs{})
 
@@ -355,7 +363,9 @@ func TestEvaluatorGetResource(t *testing.T) {
 	}
 
 	data := &evaluationStateData{
-		Evaluator: evaluator,
+		evaluationData: &evaluationData{
+			Evaluator: evaluator,
+		},
 	}
 	scope := evaluator.Scope(data, nil, nil, lang.ExternalFuncs{})
 
@@ -522,7 +532,9 @@ func TestEvaluatorGetResource_changes(t *testing.T) {
 	}
 
 	data := &evaluationStateData{
-		Evaluator: evaluator,
+		evaluationData: &evaluationData{
+			Evaluator: evaluator,
+		},
 	}
 	scope := evaluator.Scope(data, nil, nil, lang.ExternalFuncs{})
 
@@ -554,7 +566,9 @@ func TestEvaluatorGetModule(t *testing.T) {
 		cty.StringVal("bar").Mark(marks.Sensitive),
 	)
 	data := &evaluationStateData{
-		Evaluator: evaluator,
+		evaluationData: &evaluationData{
+			Evaluator: evaluator,
+		},
 	}
 	scope := evaluator.Scope(data, nil, nil, lang.ExternalFuncs{})
 	want := cty.ObjectVal(map[string]cty.Value{"out": cty.StringVal("bar").Mark(marks.Sensitive)})
